pyproject.toml
[tool.poetry]
name = "llm-eval"
version = "1.0.0"
description = "Production-grade LLM evaluation framework for RAG/QA systems with multi-metric analysis"
authors = ["LLM Eval Team"]
readme = "README.md"
license = "MIT"
packages = [{include = "llm_eval", from = "src"}]
keywords = ["llm", "evaluation", "rag", "nlp", "metrics", "bertscore", "bleu", "rouge"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.9.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.1"
pandas = "^2.1.0"
nltk = "^3.8.1"
rouge-score = "^0.1.2"
sentence-transformers = "^2.2.2"
torch = "^2.1.0"
openai = "^1.6.0"
anthropic = "^0.8.0"
groq = "^0.4.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
numpy = "^1.26.0"
rich = "^13.7.0"
tenacity = "^8.2.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.12.0"
black = "^23.12.0"
isort = "^5.13.0"
mypy = "^1.8.0"
flake8 = "^7.0.0"

[tool.poetry.scripts]
llm-eval = "llm_eval.cli:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/llm_eval"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
